Fix accounting bug
Restler uses url encoding to transmit json (if it is an object)
https://github.com/danwrong/restler/blob/e70709c7f0c998118da5a7cec82fb1a90d36d411/lib/restler.js#L58-L62

Transmitting a float url encoded would mean, the server interprets the transmitted value as a string.
eg. 250.51 -> "250.51"

Harvest interprets "250.51" as 25051 (removing the dot).
When transforming the js-object to a json string, restler sends the json directly without url encoding the data.
https://github.com/danwrong/restler/blob/e70709c7f0c998118da5a7cec82fb1a90d36d411/lib/restler.js#L63-L67

Current behavior results in huge errors when creating payments.